refactor: world with its own color palette

+++ application/src/ports/outgoing/pixel_history_store.rs
@@ -13,7 +13,7 @@ pub struct PixelHistoryEntry {
     pub username: String,
     pub pixel_x: usize,
     pub pixel_y: usize,
-    pub color_id: u8,
+    pub color_id: i16,
     pub timestamp: time::OffsetDateTime,
 }
 
@@ -21,7 +21,7 @@ pub struct PixelHistoryEntry {
 pub struct PixelInfo {
     pub user_id: Uuid,
     pub username: String,
-    pub color_id: u8,
+    pub color_id: i16,
     pub timestamp: time::OffsetDateTime,
 }
 
@@ -41,7 +41,7 @@ pub trait PixelHistoryStorePort: Send + Sync {
         &self,
         world_id: &WorldId,
         coord: TileCoord,
-    ) -> AppResult<Vec<(usize, usize, u8)>>;
+    ) -> AppResult<Vec<(usize, usize, i16)>>;
     async fn get_distinct_tile_count(&self, world_id: &WorldId, tile_size: usize)
     -> AppResult<i64>;
     async fn get_pixel_info(
